Relocate code to set retention
